+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#294] - Build option for linking against system RtAudio and RtMidi ([#263], [#288]; thanks [@OPNA2608])
 - [#308] - Support .pvi import which has a file identifier as `PVI1` (thanks [@freq-mod], [@OPNA2608])
 - [#313] - .p86, .pzi and .pps support as ADPCM samples (thanks [@OPNA2608], [@freq-mod])
+- [#323] - DRAM memory warning (thanks [@freq-mod])
 - [#286] - Polish support (thanks [@freq-mod])
 - 1 new sample module (thanks [@Zexxerd])
 
@@ -53,6 +54,7 @@
 [#304]: https://github.com/rerrahkr/BambooTracker/issues/304
 [#313]: https://github.com/rerrahkr/BambooTracker/issues/313
 [#320]: https://github.com/rerrahkr/BambooTracker/issues/320
+[#323]: https://github.com/rerrahkr/BambooTracker/issues/323
 
 ## v0.4.5 (2020-11-05)
 ### Added